% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amortization.R
\name{amort_event_table}
\alias{amort_event_table}
\title{amort_event_table}
\usage{
amort_event_table(principle, annual_interest, years, events = NULL)
}
\arguments{
\item{events}{}
}
\description{
amort_event_table
}
