% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amortization.R
\name{create_table}
\alias{create_table}
\title{create_table}
\usage{
create_table(payment, start_principle, annual_interest, month_pay,
  n_payments)
}
\arguments{
\item{n_payments}{}
}
\description{
create_table
}
