% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amortization.R
\name{length_mortgage}
\alias{length_mortgage}
\title{length_mortgage}
\usage{
length_mortgage(principle, annual_interest, month_payment)
}
\arguments{
\item{month_payment}{}
}
\description{
length_mortgage
}
